requirements.txt
aiohttp
opal-common
pydantic
python-dotenv
